+++ src/test/java/roomescape/unit/service/reservationmember/ReserveTicketServiceTest.java
@@ -25,7 +25,7 @@ import roomescape.service.theme.ThemeService;
 import roomescape.unit.config.ServiceFixture;
 import roomescape.unit.repository.member.FakeMemberRepository;
 import roomescape.unit.repository.reservation.FakeReservationRepository;
-import roomescape.unit.repository.reservation.FakeReservationTimeRepository;
+import roomescape.unit.repository.reservationtime.FakeReservationTimeRepository;
 import roomescape.unit.repository.reservation.FakeThemeRepository;
 import roomescape.unit.repository.reserveticket.FakeReserveTicketRepository;
